requirements.txt
catalyst==21.05
numpy>=1.17.4
opencv-python>=4.0
torch==1.8.0
albumentations==0.4.3
torch_optimizer==0.1.0
mlflow==1.14.0
scikit-multilearn==0.2.0
torchvision==0.9.0
torch-tools==0.1.5
onnxruntime==1.7.0
scikit-learn==0.24.1
onnxruntime==1.7.0
ml-pyxis
mlconfig==0.1.0
flake8
pandas==1.1.5
python-dotenv==0.19.1
boto3==1.19.3
botocore==1.22.3
tqdm==4.62.3
typing==3.7.4.3
PyYAML==6.0
split-folders==0.4.3
dvc[s3]==2.1.0
Pillow==8.4.0
tensorflow==2.5.0
aiohttp
pandarallel==1.5.4
